scripts/institution-bot/requirements.txt
google-adk
google-generativeai
PyGithub
requests
beautifulsoup4
python-dotenv
tavily-python
